Crate alloy_rpc_types_eth

alloy-rpc-types-eth

Types for the eth Ethereum JSON-RPC namespace.

Re-exports

pub use erc4337::PackedUserOperation;

pub use erc4337::SendUserOperation;

pub use erc4337::SendUserOperationResponse;

pub use erc4337::UserOperation;

pub use erc4337::UserOperationGasEstimation;

pub use erc4337::UserOperationReceipt;

pub use transaction::*;

Modules

erc4337

This module provides implementations for EIP-4337.

error

Commonly used errors for the eth_ namespace.

pubsub

Ethereum types for pub-sub

simulate

‘eth_simulateV1’ Request / Response types: https://github.com/ethereum/execution-apis/pull/484

state

bindings for state overrides in eth_call

transaction

RPC types for transactions

Structs

Account

Represents an TrieAccount in the account trie.

AccountInfo

Account information.

BadBlock

Bad block representation.

Block

Block representation for RPC.

BlockOverrides

BlockOverrides is a set of header fields to override.

BlockTransactionHashes

An iterator over the transaction hashes of a block.

BloomFilter

Helper type to represent a bloom filter used for matching logs.

Bundle

Bundle of transactions

ChainStatus

Chain status.

EIP1186AccountProofResponse

Response for EIP-1186 account proof eth_getProof

EIP1186StorageProof

Data structure with proof for one single storage-entry

EthCallResponse

CallResponse for eth_callMany

ExtAccountInfo

Extended account information (used by parity_allAccountInfo).

FeeHistory

Response type for eth_feeHistory

Filter

Filter for logs.

FilterSet

FilterSet is a set of values that will be used to filter logs.

FilteredParams

Support for matching Filters

Header

RPC representation of block header, wrapping a consensus header.

Index

A hex encoded or decimal index that’s intended to be used as a rust index, hence it’s deserialized into a usize.

Log

Ethereum Log emitted by a transaction

PeerEthProtocolInfo

Peer Ethereum protocol information

PeerInfo

Peer connection information

PeerNetworkInfo

Peer network information

PeerProtocolsInfo

Peer protocols information

Peers

Peers info

PipProtocolInfo

Peer PIP protocol information

RawLog

Ethereum Log

RecoveredAccount

account derived from a signature as well as information that tells if it is valid for the current chain

RpcBlockHash

A block hash which may have a boolean requireCanonical field.

Stage

The detail of the sync stages.

StateContext

State context for callMany

SyncInfo

Syncing info

TransactionStats

Propagation statistics for pending transaction.

TxGasAndReward

Internal struct to calculate reward percentiles

Withdrawal

Withdrawal represents a validator withdrawal from the consensus layer.

Withdrawals

Represents a collection of Withdrawals.

Work

The result of an eth_getWork request

Enums

BlockError

Error that can occur when converting other types to blocks

BlockId

A Block Identifier. https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md

BlockNumberOrTag

A block Number (or tag - “latest”, “earliest”, “pending”)

BlockTransactions

Block Transactions depending on the boolean attribute of eth_getBlockBy*, or if used by eth_getUncle*

BlockTransactionsKind

Determines how the transactions field of block should be filled.

FilterBlockError

Represents errors that can occur when setting block filters in FilterBlockOption.

FilterBlockOption

Represents the target range of blocks for the filter

FilterChanges

Response of the eth_getFilterChanges RPC.

FilterId

Owned equivalent of a SubscriptionId

PendingTransactionFilterKind

Specifies the kind of information you wish to receive from the eth_newPendingTransactionFilter RPC endpoint.

SyncStatus

Sync status

TransactionIndex

Represents a transaction index where -1 means all transactions

ValueOrArray

Union type for representing a single value or a vector of values inside a filter

Functions

calc_blob_gasprice

Calculates the blob gas price from the header’s excess blob gas field.

calc_excess_blob_gas

Calculates the excess_blob_gas from the parent header’s blob_gas_used and excess_blob_gas.

logs_bloom

Calculate receipt logs bloom.

Type Aliases

BlockHashOrNumber

A block hash or a block number

BlockNumHash

A block number and a hash

ForkBlock

Block number and hash of the forked block.

Topic

A single topic