alloy_rpc_types_eth::state

Struct AccountOverride

pub struct AccountOverride {
    pub balance: Option<U256>,
    pub nonce: Option<u64>,
    pub code: Option<Bytes>,
    pub state: Option<B256HashMap<B256>>,
    pub state_diff: Option<B256HashMap<B256>>,
    pub move_precompile_to: Option<Address>,
}

Custom account override used in call

Fields
balance: Option<U256>

Fake balance to set for the account before executing the call.

nonce: Option<u64>

Fake nonce to set for the account before executing the call.

code: Option<Bytes>

Fake EVM bytecode to inject into the account before executing the call.

state: Option<B256HashMap<B256>>

Fake key-value mapping to override all slots in the account storage before executing the call.

state_diff: Option<B256HashMap<B256>>

Fake key-value mapping to override individual slots in the account storage before executing the call.

move_precompile_to: Option<Address>

Moves addresses precompile into the specified address. This move is done before the โcodeโ override is set. When the specified address is not a precompile, the behaviour is undefined and different clients might behave differently.

Implementations

impl AccountOverride

pub fn with_code(self, code: impl Into<Bytes>) -> Self

Configures the bytecode override

pub fn with_state(self, state: impl IntoIterator<Item = (B256, B256)>) -> Self

Configures the state overrides

pub fn with_state_diff( self, state_diff: impl IntoIterator<Item = (B256, B256)>, ) -> Self

Configures the state diffs

pub const fn with_balance(self, balance: U256) -> Self

Configures the balance override

pub const fn with_nonce(self, nonce: u64) -> Self

Configures the nonce override

pub fn set_code(&mut self, code: impl Into<Bytes>)

Sets the bytecode override in place.

pub fn set_state(&mut self, state: impl IntoIterator<Item = (B256, B256)>)

Sets the state overrides in place.

pub fn set_state_diff( &mut self, state_diff: impl IntoIterator<Item = (B256, B256)>, )

Sets the state diffs in place.

pub const fn set_balance(&mut self, balance: U256)

Sets the balance override in place.

pub const fn set_nonce(&mut self, nonce: u64)

Sets the nonce override in place.

pub const fn set_move_precompile_to(&mut self, address: Address)

Sets the move precompile address in place.

pub fn with_code_opt(self, code: Option<impl Into<Bytes>>) -> Self

Conditionally sets the bytecode override and returns self.

pub const fn with_balance_opt(self, balance: Option<U256>) -> Self

Conditionally sets the balance override and returns self.

pub const fn with_nonce_opt(self, nonce: Option<u64>) -> Self

Conditionally sets the nonce override and returns self.

pub const fn with_move_precompile_to_opt(self, address: Option<Address>) -> Self

Conditionally sets the move precompile address and returns self.