alloy_rpc_types_eth

Struct Block

pub struct Block<T = Transaction<TxEnvelope>, H = Header> {
    pub header: H,
    pub uncles: Vec<B256>,
    pub transactions: BlockTransactions<T>,
    pub withdrawals: Option<Withdrawals>,
}

Block representation for RPC.

Fields
header: H

Header of the block.

uncles: Vec<B256>

Uncles’ hashes.

transactions: BlockTransactions<T>

Block Transactions. In the case of an uncle block, this field is not included in RPC responses, and when deserialized, it will be set to BlockTransactions::Uncle.

withdrawals: Option<Withdrawals>

Withdrawals in the block.

Implementations

impl<T, H> Block<T, H>

pub const fn empty(header: H) -> Self

Creates a new empty block (without transactions).

pub const fn new(header: H, transactions: BlockTransactions<T>) -> Self

Creates a new Block with the given header and transactions.

Note: This does not set the withdrawals for the block.

use alloy_eips::eip4895::Withdrawals;
use alloy_network_primitives::BlockTransactions;
use alloy_rpc_types_eth::{Block, Header, Transaction};
let block = Block::new(
    Header::new(alloy_consensus::Header::default()),
    BlockTransactions::<Transaction>::Full(vec![]),
)
.with_withdrawals(Some(Withdrawals::default()));

pub fn number(&self) -> u64
where H: BlockHeader,

Returns the block’s number.

pub fn apply<F>(self, f: F) -> Self
where F: FnOnce(Self) -> Self,

Apply a function to the block, returning the modified block.

pub fn with_transactions(self, transactions: BlockTransactions<T>) -> Self

Sets the transactions for the block.

pub fn with_withdrawals(self, withdrawals: Option<Withdrawals>) -> Self

Sets the withdrawals for the block.

pub fn with_uncles(self, uncles: Vec<B256>) -> Self

Sets the uncles for the block.

pub fn try_into_transactions( self, ) -> Result<Vec<T>, ValueError<BlockTransactions<T>>>

Tries to convert inner transactions into a vector of full transactions

Returns an error if the block contains only transaction hashes or if it is an uncle block.

pub fn into_transactions_vec(self) -> Vec<T>

Consumes the type and returns the transactions as a vector.

Note: if this is an uncle or hashes, this will return an empty vector.

pub fn try_into_block_body(self) -> Result<BlockBody<T, H>, ValueError<Self>>

Converts this block into a BlockBody.

Returns an error if the transactions are not full or if the block has uncles.

pub fn into_block_body_unchecked(self) -> BlockBody<T, H>

Converts this block into a BlockBody

Caution: The body will have empty transactions unless the block’s transactions are BlockTransactions::Full. This will disregard ommers/uncles and always return an empty ommers vec.

pub fn into_consensus_block(self) -> Block<T, H>

Consumes the block and returns the alloy_consensus::Block with the current transaction and header type.

Note: Unlike Self::into_consensus, this method returns the Header type H as-is without converting it to alloy_consensus::Header, See Header::into_consensus.

This has two caveats:

The returned block will always have empty uncles.

If the block’s transaction is not BlockTransactions::Full, the returned block will have an empty transaction vec.

pub fn map_header<U>(self, f: impl FnOnce(H) -> U) -> Block<T, U>

Converts the block’s header type by applying a function to it.

pub fn into_header(self) -> H

Consumes the block and only returns the rpc header.

To obtain the underlying alloy_consensus::Header use Block::into_consensus_header.

pub fn try_convert_header<U>(self) -> Result<Block<T, U>, U::Error>
where U: TryFrom<H>,

Converts the block’s header type to the given alternative that is TryFrom<H>

pub fn try_map_header<U, E>( self, f: impl FnOnce(H) -> Result<U, E>, ) -> Result<Block<T, U>, E>

Converts the block’s header type by applying a fallible function to it.

pub fn convert_transactions<U>(self) -> Block<U, H>
where U: From<T>,

Converts the block’s transaction type to the given alternative that is From<T>

pub fn try_convert_transactions<U>(self) -> Result<Block<U, H>, U::Error>
where U: TryFrom<T>,

Converts the block’s transaction to the given alternative that is TryFrom<T>

Returns the block with the new transaction type if all conversions were successful.

pub fn map_transactions<U>(self, f: impl FnMut(T) -> U) -> Block<U, H>

Converts the block’s transaction type by applying a function to each transaction.

Returns the block with the new transaction type.

pub fn try_map_transactions<U, E>( self, f: impl FnMut(T) -> Result<U, E>, ) -> Result<Block<U, H>, E>

Converts the block’s transaction type by applying a fallible function to each transaction.

Returns the block with the new transaction type if all transactions were mapped successfully.

pub fn calculate_transactions_root(&self) -> Option<B256>
where T: Encodable2718,

Calculate the transaction root for the full transactions in this block type.

Returns None if the transactions is not the BlockTransactions::Full variant.

impl<T: TransactionResponse, H> Block<T, H>

pub fn into_full_block(self, txs: Vec<T>) -> Self

Converts a block with Tx hashes into a full block.

impl<T, H: Sealable + Encodable> Block<T, Header<H>>

pub fn uncle_from_header(header: H) -> Self

Constructs an “uncle block” from the provided header.

This function creates a new Block structure for uncle blocks (ommer blocks), using the provided alloy_consensus::Header.

impl<T> Block<T>

pub const fn hash(&self) -> B256

Returns the block’s hash as received from rpc.

pub const fn sealed_header(&self) -> Sealed<&Header>

Returns a sealed reference of the header: Sealed<&Header>

pub fn into_sealed_header(self) -> Sealed<Header>

Consumes the type and returns the sealed alloy_consensus::Header.

pub fn into_consensus_header(self) -> Header

Consumes the type, strips away the rpc context from the rpc Header type and just returns the alloy_consensus::Header.

pub fn from_consensus(block: Block<T>, total_difficulty: Option<U256>) -> Self
where T: Encodable,

Constructs block from a consensus block and total_difficulty.

pub fn into_consensus(self) -> Block<T>

Consumes the block and returns the ethereum alloy_consensus::Block with the ethereum header type.

This has two caveats:

The returned block will always have empty uncles.

If the block’s transaction is not BlockTransactions::Full, the returned block will have an empty transaction vec.

pub fn into_consensus_sealed(self) -> Sealed<Block<T>>

Same as Self::into_consensus but returns the block as Sealed with the block’s hash.