alloy_rpc_types_eth

Struct Log

pub struct Log<T = LogData> {
    pub inner: Log<T>,
    pub block_hash: Option<BlockHash>,
    pub block_number: Option<u64>,
    pub block_timestamp: Option<u64>,
    pub transaction_hash: Option<TxHash>,
    pub transaction_index: Option<u64>,
    pub log_index: Option<u64>,
    pub removed: bool,
}

Ethereum Log emitted by a transaction

Fields
inner: Log<T>

Consensus log object

block_hash: Option<BlockHash>

Hash of the block the transaction that emitted this log was mined in

block_number: Option<u64>

Number of the block the transaction that emitted this log was mined in

block_timestamp: Option<u64>

The timestamp of the block as proposed in: https://ethereum-magicians.org/t/proposal-for-adding-blocktimestamp-to-logs-object-returned-by-eth-getlogs-and-related-requests https://github.com/ethereum/execution-apis/issues/295

transaction_hash: Option<TxHash>

Transaction Hash

transaction_index: Option<u64>

Index of the Transaction in the block

log_index: Option<u64>

Log Index in Block

removed: bool

Geth Compatibility Field: whether this log was removed

Implementations

impl<T> Log<T>

pub const fn address(&self) -> Address

Getter for the address field. Shortcut for log.inner.address.

pub const fn data(&self) -> &T

Getter for the data field. Shortcut for log.inner.data.

pub fn into_inner(self) -> Log<T>

Consumes the type and returns the wrapped alloy_primitives::Log

impl Log<LogData>

pub fn topics(&self) -> &[B256]

Getter for the topics field. Shortcut for log.inner.topics().

pub fn topic0(&self) -> Option<&B256>

Getter for the topic0 field.

pub fn topics_mut(&mut self) -> &mut [B256]

Get the topic list, mutably. This gives access to the internal array, without allowing extension of that array. Shortcut for LogData::topics_mut

pub fn log_decode<T: SolEvent>(&self) -> Result<Log<T>>

Decode the log data into a typed log.

pub fn log_decode_validate<T: SolEvent>(&self) -> Result<Log<T>>

Decode the log data with validation into a typed log.

pub fn collect_for_receipt<I, T>( previous_log_count: usize, meta: TransactionMeta, logs: I, ) -> Vec<Log<T>>
where I: IntoIterator<Item = Log<T>>,

Creates a collection of RPC logs from transaction receipt logs.

This function takes raw consensus logs and enriches them with RPC metadata needed for API responses, including block information and proper indexing.

Arguments

previous_log_count - The total number of logs from previous transactions in the same block. Used to calculate the correct log_index for each log.

meta - Transaction metadata containing block hash, number, timestamp, and transaction information needed to populate the RPC log fields.

logs - An iterator of consensus logs to be converted into RPC logs.

Returns

A vector of RPC logs with all metadata fields populated, ready to be included in the transaction receipt.

impl<T> Log<T>
where for<'a> &'a T: Into<LogData>,

pub fn reserialize_inner(&self) -> Log

Reserialize the inner data, returning an alloy_primitives::Log.

pub fn reserialize(&self) -> Log<LogData>

Reserialize the data, returning a new Log object wrapping an alloy_primitives::Log. this copies the log metadata, preserving the original object.