Module alloy_rpc_types_eth::transaction

RPC types for transactions

Re-exports

pub use optimism::OptimismTransactionReceiptFields;

pub use request::TransactionInput;

pub use request::TransactionRequest;

Modules

common 🔒

Commonly used additional types that are not part of the JSON RPC spec but are often required when working with RPC types, such as Transaction

error 🔒

optimism

Misc Optimism-specific types.

receipt 🔒

request

Alloy basic Transaction Request type.

signature 🔒

Signature related RPC values

Structs

AccessList

AccessList as defined in EIP-2930

AccessListItem

A list of addresses and storage keys that the transaction plans to access. Accesses outside the list are possible, but become more expensive.

AccessListWithGasUsed

Access list with gas used appended.

AnyReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

Parity

Type that represents the signature parity byte, meant for use in RPC.

Receipt

Receipt containing result of transaction execution.

ReceiptWithBloom

Receipt with calculated bloom filter.

Signature

Container type for all signature fields in RPC

Transaction

Transaction object used in RPC

TransactionInfo

Additional fields in the context of a block that contains this transaction.

TransactionReceipt

Transaction receipt

Enums

ConversionError

Error variants when converting from crate::Transaction to [alloy_consensus::Signed] transaction.

ReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

Type Aliases

AnyTransactionReceipt

Alias for a catch-all receipt type.