alloy_rpc_types_eth::transaction

Module request

Alloy basic Transaction Request type.

Structs

BuildTransactionErr

Error thrown when a transaction request cannot be built into a transaction.

TransactionInput

Helper type that supports both data and input fields that map to transaction input data.

TransactionInputError

Error thrown when both data and input fields are set and not equal.

TransactionRequest

Represents all transaction requests to/from RPC.

Enums

TransactionInputKind

Represents how a TransactionRequest handles input/data fields.