alloy_rpc_types_eth::transaction
Struct Transaction

pub struct Transaction {

    pub hash: TxHash,
    pub nonce: u64,
    pub block_hash: Option<BlockHash>,
    pub block_number: Option<u64>,
    pub transaction_index: Option<u64>,
    pub from: Address,
    pub to: Option<Address>,
    pub value: U256,
    pub gas_price: Option<u128>,
    pub gas: u64,
    pub max_fee_per_gas: Option<u128>,
    pub max_priority_fee_per_gas: Option<u128>,
    pub max_fee_per_blob_gas: Option<u128>,
    pub input: Bytes,
    pub signature: Option<Signature>,
    pub chain_id: Option<ChainId>,
    pub blob_versioned_hashes: Option<Vec<B256>>,
    pub access_list: Option<AccessList>,
    pub transaction_type: Option<u8>,
    pub authorization_list: Option<Vec<SignedAuthorization>>,

}

Transaction object used in RPC

Fields
hash: TxHash

Hash

nonce: u64

Nonce

block_hash: Option<BlockHash>

Block hash

block_number: Option<u64>

Block number

transaction_index: Option<u64>

Transaction Index

from: Address

Sender

to: Option<Address>

Recipient

value: U256

Transferred value

gas_price: Option<u128>

Gas Price

gas: u64

Gas amount

max_fee_per_gas: Option<u128>

Max BaseFeePerGas the user is willing to pay.

max_priority_fee_per_gas: Option<u128>

The miner’s tip.

max_fee_per_blob_gas: Option<u128>

Configured max fee per blob gas for eip-4844 transactions

input: Bytes

Data

signature: Option<Signature>

All flattened fields of the transaction signature.

Note: this is an option so special transaction types without a signature (e.g. https://github.com/ethereum-optimism/optimism/blob/0bf643c4147b43cd6f25a759d331ef3a2a61a2a3/specs/deposits.md#the-deposited-transaction-type) can be supported.

chain_id: Option<ChainId>

The chain id of the transaction, if any.

blob_versioned_hashes: Option<Vec<B256>>

Contains the blob hashes for eip-4844 transactions.

access_list: Option<AccessList>

EIP2930

Pre-pay to warm storage access.

transaction_type: Option<u8>

EIP2718

Transaction type, Some(4) for EIP-7702 transaction, Some(3) for EIP-4844 transaction, Some(2) for EIP-1559 transaction, Some(1) for AccessList transaction, None or Some(0) for Legacy

authorization_list: Option<Vec<SignedAuthorization>>

The signed authorization list is a list of tuples that store the address to code which the signer desires to execute in the context of their EOA and their signature.

Implementations

impl Transaction

pub const fn is_legacy_gas(&self) -> bool

Returns true if the transaction is a legacy or 2930 transaction.

pub fn into_request(self) -> TransactionRequest

Converts Transaction into TransactionRequest.

During this conversion data for TransactionRequest::sidecar is not populated as it is not part of Transaction.