Struct alloy_rpc_types_eth::Header

pub struct Header<H = Header> {
    pub hash: BlockHash,
    pub inner: H,
    pub total_difficulty: Option<U256>,
    pub size: Option<U256>,
}

RPC representation of block header, wrapping a consensus header.

This wraps the consensus header and adds additional fields for RPC.

Fields
hash: BlockHash

Hash of the block

inner: H

Inner consensus header.

total_difficulty: Option<U256>

Total difficulty

Note: This field is now effectively deprecated: https://github.com/ethereum/execution-apis/pull/570

size: Option<U256>

Integer the size of this block in bytes.

Implementations

impl<H> Header<H>

pub fn new(inner: H) -> Self
where H: Sealable,

Create a new Header from a consensus header.

Note: This will compute the hash of the header.

pub fn from_sealed(header: Sealed<H>) -> Self

Create a new Header from a sealed consensus header.

Note: This does not set the total difficulty or size of the block.

pub fn into_consensus(self) -> H

Consumes the type and returns the wrapped consensus header.

pub fn from_consensus( header: Sealed<H>, total_difficulty: Option<U256>, size: Option<U256>, ) -> Self

Create a new Header from a sealed consensus header and additional fields.

pub const fn with_total_difficulty(self, total_difficulty: Option<U256>) -> Self

Set the total difficulty of the block.

pub const fn with_size(self, size: Option<U256>) -> Self

Set the size of the block.

impl<H: BlockHeader> Header<H>

pub fn blob_fee(&self) -> Option<u128>

Returns the blob fee for this block according to the EIP-4844 spec.

Returns None if excess_blob_gas is None

pub fn next_block_blob_fee(&self, blob_params: BlobParams) -> Option<u128>

Returns the blob fee for the next block according to the EIP-4844 spec.

Returns None if excess_blob_gas is None.

See also Self::next_block_excess_blob_gas

pub fn next_block_excess_blob_gas(&self, blob_params: BlobParams) -> Option<u64>

Calculate excess blob gas for the next block according to the EIP-4844 spec.

Returns a None if no excess blob gas is set, no EIP-4844 support