alloy_rpc_types_eth

Module transaction

RPC types for transactions

Re-exports

pub use request::TransactionInput;

pub use request::TransactionInputKind;

pub use request::TransactionRequest;

Modules

request

Alloy basic Transaction Request type.

Structs

AccessList

AccessList as defined in EIP-2930

AccessListItem

A list of addresses and storage keys that the transaction plans to access. Accesses outside the list are possible, but become more expensive.

AccessListResult

AccessListResult for handling errors from eth_createAccessList

AnyReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

Authorization

An unsigned EIP-7702 authorization.

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

Receipt

Receipt containing result of transaction execution.

ReceiptWithBloom

Receipt with calculated bloom filter.

Transaction

Transaction object used in RPC.

TransactionInfo

Additional fields in the context of a (maybe) pending block that contains this transaction.

TransactionReceipt

Transaction receipt

Enums

ConversionError

Error variants when converting from crate::Transaction to alloy_consensus::Signed transaction.

ReceiptEnvelope

Receipt envelope, as defined in EIP-2718.

Traits

TransactionTrait

Represents a minimal EVM transaction. Currently, EIP-1559, EIP-4844, and EIP-7702 support dynamic fees. We call these transactions “dynamic fee transactions”. We call non dynamic fee transactions(EIP-155, EIP-2930) “legacy fee transactions”.